requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
statsmodels==0.14.2
scikit-learn==1.4.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
python-dateutil==2.9.0.post0
requests==2.32.3
streamlit==1.37.1
openai==1.40.0
plotly==6.3.0
python-dotenv==1.1.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.21